Refactor API headers
